requirements/base.txt
django==5.2.5
django-environ==0.12.0
django-model-utils==5.0.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework-simplejwt==5.5.1
django-mptt==0.17.0
django-simple-history==3.10.1
